tasks_CRUD_edit

+++ app/Http/Controllers/TasksController.php
@@ -14,12 +14,19 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $tasks = Task::orderBy('id', 'desc')->paginate(25);
+    {   
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(25);
+       
+          $data = [
+              'user' => $user,
+              'tasks' => $tasks,
+              ];
+        }
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        return view('welcome', $data);
     }
 
     /**
@@ -30,10 +37,9 @@ class TasksController extends Controller
     public function create()
     {
         $task = new Task;
-        
         return view('tasks.create', [
             'task' => $task,
-            ]);
+        ]);
     }
 
     /**
@@ -49,13 +55,21 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required',
         ]);
-            
+        
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = \Auth::id();
         $task->save();
         
         return redirect('/');
+        /*
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            ]);
+         */   
+    
     }
 
     /**
@@ -68,6 +82,9 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
+       if(\Auth::id() !== $task->user_id){
+         return redirect('/');  
+       } 
         return view('tasks.show',[
             'task' => $task,
             ]);
@@ -80,12 +97,30 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
+    {    
         $task = Task::find($id);
         
-        return view('tasks.edit', [
-            'task' => $task,
-            ]);
+        if (\Auth::id() !== $task->user_id){
+            
+            return redirect('/');
+        }
+        
+             return view('tasks.edit', [
+                'task' => $task,
+             ]);
+            
+               
+        
+        
+        // if (\Auth::id() === $task->user_id){
+            
+        //     return view('tasks.edit', [
+        //         'task' => $task,
+        //     ]);
+            
+        // }else{
+        //     return redirect('/');
+        // }
     }
 
     /**
@@ -97,18 +132,35 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $task = Task::find($id);
+        
+        if(\Auth::id() !==$task->user_id){
+            return redirect('/'); 
+        }
+        
         //バリデーション
         $this->validate($request, [
             'status' => 'required|max:10',
             'content' => 'required',
             ]);
             
-        $task = Task::find($id);
+        
         $task ->status = $request->status;
         $task ->content = $request->content;
+        $task ->user_id = \Auth::id();
         $task->save();
         
         return redirect('/');
+        
+        
+        /*
+         $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            ]);
+        */
+        
+        
     }
 
     /**
@@ -120,8 +172,15 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
+        if(\Auth::id() !==$task->user_id){
+            return redirect('/');
+    
+        } 
+        
+         $task->delete();
         return redirect('/');
+        
+        
     }
 }
